+++ spec/integration/braintree/transaction_gateway_us_bank_account_spec.js
@@ -118,7 +118,7 @@ describe('TransactionGateway', function () {
       privateKey: 'integration2_private_key'
     };
 
-    let gateway = braintree.connect(config);
+    let gateway = new braintree.BraintreeGateway(config);
 
     context('plaid verified', function () {
       it('transacts successfully', function (done) {