+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Add `bin` field to `paymentMethodNonceDetails`
 * Add Error indicating pdf uploads too long for dispute evidence.
 * Add `GrantedPaymentMethodRevoked` webhook response objects
+* Add ability to instantiate a new Braintree Gateway directly
 
 ## 2.14.0
 * Add `processor_response_type` to `Transaction`, `AuthorizationAdjustment`, and `CreditCardVerification`.